Remove unused header
